+++ packages/tui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.85.1] - 2026-09-05
+
+### Added
+
+- Added five-times-faster mouse wheel scrolling while holding Alt in fullscreen mode ([#9166](https://github.com/earendil-works/pi/pull/9166) by [@xl0](https://github.com/xl0)).
+
+### Fixed
+
+- Fixed mouse hover changing selection and recentering autocomplete and settings lists, causing clicks to target a different item.
+
 ## [0.85.0] - 2026-09-04
 
 ### Breaking Changes